Move testables to dedicated packages
