+++ modules/pipeline/jimple-library-usage-graph-generator/src/test/java/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/testclasses/users/ifconditional/IfBothUseTest.java
@@ -1,15 +1,15 @@
-package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users;
+package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users.ifconditional;
 
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.library.Object1;
 
-public class IfTrueUseTest {
+public class IfBothUseTest {
     public Object1 test() {
         Object1 o1 = new Object1();
 
         if (Math.random() > 123) {
-            o1.m3();
+            o1.m1();
         } else {
-            new String("Cool input string");
+            o1.m3();
         }
 
         return o1;
