+++ modules/pipeline/jimple-library-usage-graph-generator/src/test/java/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/testclasses/users/ifconditional/IfFalseUseTest.java
@@ -1,15 +1,15 @@
-package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users;
+package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users.ifconditional;
 
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.library.Object1;
 
-public class IfNoUseTest {
+public class IfFalseUseTest {
     public Object1 test() {
         Object1 o1 = new Object1();
 
         if (Math.random() > 123) {
-            new String("Cool input string 1");
+            new String("Cool input string");
         } else {
-            new String("Cool input string 2");
+            o1.m1();
         }
 
         return o1;
