+++ modules/pipeline/jimple-library-usage-graph-generator/src/test/java/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/testclasses/users/switchconditional/SwitchDefaultUseTest.java
@@ -1,9 +1,9 @@
-package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users;
+package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users.switchconditional;
 
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.library.Object1;
 
-public class SwitchOneUseTest {
-    public void test() {
+public class SwitchDefaultUseTest {
+    public Object1 test() {
         Object1 o1 = new Object1();
 
         switch (o1.hashCode()) {
@@ -12,11 +12,11 @@ public class SwitchOneUseTest {
                 break;
             case 2:
                 int b = 2;
-                o1.m1();
                 break;
-            case 3:
-                int c = 3;
-                break;
+            default:
+                o1.m3();
         }
+
+        return o1;
     }
 }
