+++ modules/pipeline/jimple-library-usage-graph-generator/src/test/java/org/cafejojo/schaapi/pipeline/usagegraphgenerator/jimple/testclasses/users/switchconditional/SwitchNoUseTest.java
@@ -1,4 +1,4 @@
-package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users;
+package org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.users.switchconditional;
 
 import org.cafejojo.schaapi.pipeline.usagegraphgenerator.jimple.testclasses.library.Object1;
